"DumpFile template class for all pipelines; updates #107"
