+++ prototype2/common/DumpFileTest.cpp
@@ -9,32 +9,32 @@ class HitFileTest : public TestBase {
 protected:
   virtual void SetUp() {
     hdf5::error::Singleton::instance().auto_print(false);
-    if (boost::filesystem::exists("hit_file_test.h5"))
+    if (boost::filesystem::exists("readout_file_test.h5"))
     {
-      boost::filesystem::remove("hit_file_test.h5");
+      boost::filesystem::remove("readout_file_test.h5");
     }
 
-    if (boost::filesystem::exists("hit_file_test_1.h5"))
+    if (boost::filesystem::exists("readout_file_test_1.h5"))
     {
-      boost::filesystem::remove("hit_file_test_1.h5");
+      boost::filesystem::remove("readout_file_test_1.h5");
     }
   }
   virtual void TearDown() {}
 };
 
 TEST_F(HitFileTest, CreateFile) {
-  HitFile::create("hit_file_test");
-  EXPECT_TRUE(hdf5::file::is_hdf5_file("hit_file_test.h5"));
+  HitFile::create("readout_file_test");
+  EXPECT_TRUE(hdf5::file::is_hdf5_file("readout_file_test.h5"));
 }
 
 TEST_F(HitFileTest, OpenEmptyFile) {
-  HitFile::create("hit_file_test");
-  auto file = HitFile::open("hit_file_test");
+  HitFile::create("readout_file_test");
+  auto file = HitFile::open("readout_file_test");
   EXPECT_EQ(file->count(), 0);
 }
 
 TEST_F(HitFileTest, Push) {
-  auto file = HitFile::create("hit_file_test");
+  auto file = HitFile::create("readout_file_test");
   file->push(std::vector<Hit>(100, Hit()));
   EXPECT_EQ(file->count(), 0);
   file->push(std::vector<Hit>(900, Hit()));
@@ -44,7 +44,7 @@ TEST_F(HitFileTest, Push) {
 }
 
 TEST_F(HitFileTest, PushFileRotation) {
-  auto file = HitFile::create("hit_file_test", 1);
+  auto file = HitFile::create("readout_file_test", 1);
   file->push(std::vector<Hit>(100, Hit()));
   EXPECT_EQ(file->count(), 0);
   file->push(std::vector<Hit>(900, Hit()));
@@ -52,19 +52,19 @@ TEST_F(HitFileTest, PushFileRotation) {
   file->push(std::vector<Hit>(3000, Hit()));
   EXPECT_EQ(file->count(), 4000);
 
-  EXPECT_TRUE(hdf5::file::is_hdf5_file("hit_file_test.h5"));
+  EXPECT_TRUE(hdf5::file::is_hdf5_file("readout_file_test.h5"));
 
-  EXPECT_FALSE(boost::filesystem::exists("hit_file_test_1.h5"));
+  EXPECT_FALSE(boost::filesystem::exists("readout_file_test_1.h5"));
   file->push(std::vector<Hit>(300000, Hit()));
-  EXPECT_TRUE(hdf5::file::is_hdf5_file("hit_file_test_1.h5"));
+  EXPECT_TRUE(hdf5::file::is_hdf5_file("readout_file_test_1.h5"));
 }
 
 TEST_F(HitFileTest, Read) {
-  auto file_out = HitFile::create("hit_file_test");
+  auto file_out = HitFile::create("readout_file_test");
   file_out->push(std::vector<Hit>(900, Hit()));
   file_out.reset();
 
-  auto file = HitFile::open("hit_file_test");
+  auto file = HitFile::open("readout_file_test");
   EXPECT_EQ(file->Data.size(), 0);
   file->readAt(0, 3);
   EXPECT_EQ(file->Data.size(), 3);
@@ -75,12 +75,12 @@ TEST_F(HitFileTest, Read) {
 }
 
 TEST_F(HitFileTest, ReadAll) {
-  auto file_out = HitFile::create("hit_file_test");
+  auto file_out = HitFile::create("readout_file_test");
   file_out->push(std::vector<Hit>(900, Hit()));
   file_out.reset();
 
   std::vector<Hit> data;
-  HitFile::read("hit_file_test", data);
+  HitFile::read("readout_file_test", data);
   EXPECT_EQ(data.size(), 900);
 }
