+++ prototype2/gdgem/clustering/HitsQueueTest.cpp
@@ -26,7 +26,7 @@ protected:
 
   virtual void SetUp() {
     std::string DataPath = TEST_DATA_PATH;
-    ReadoutFile::read(DataPath + "run16long.h5", long_data);
+    ReadoutFile::read(DataPath + "run16long", long_data);
 
     srstime.set_bc_clock(20);
     srstime.set_tac_slope(60);
