+++ prototype2/gdgem/generators/CMakeLists.txt
@@ -16,7 +16,6 @@ set(nmxgen_apv_INC
 create_executable(nmxgen_apv)
 
 set(BuilderAPVTest_SRC
-    ../nmx/HitFile.cpp
     ../nmx/Event.cpp
     ../nmx/Cluster.cpp
     ../nmx/Hists.cpp
@@ -39,7 +38,6 @@ set(BuilderAPVTest_INC
 #=============================================================================
 
 set(nmxgen_hits_SRC
-    ../nmx/HitFile.cpp
     ../nmx/Hit.cpp
     nmxgen_hits.cpp
     ReaderHits.cpp
