+++ prototype2/gdgem/nmx/ReadoutFile.h
@@ -8,40 +8,60 @@
 #pragma once
 
 #include <gdgem/nmx/Readout.h>
-#include <vector>
+#include <common/DumpFile.h>
 
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wall"
-#pragma GCC diagnostic ignored "-Wpedantic"
-#pragma GCC diagnostic ignored "-Wunused-variable"
-#pragma GCC diagnostic ignored "-Wunused-parameter"
-#include <h5cpp/hdf5.hpp>
-#pragma GCC diagnostic pop
 
-class ReadoutFile
+namespace hdf5 {
+
+namespace datatype {
+template<>
+class TypeTrait<Readout>
 {
-public:
-  ReadoutFile();
-
-  void open_rw(boost::filesystem::path file_path);
-  void open_r(boost::filesystem::path file_path);
-
-  static ReadoutFile create(boost::filesystem::path file_path);
-  static ReadoutFile open(boost::filesystem::path file_path);
-
-  size_t count() const;
-  void write();
-  void read_at(size_t idx, size_t count);
-
-  static void read(std::string file, std::vector<Readout>& external_data);
-
-  std::vector<Readout> data;
-
-  static constexpr size_t chunk_size{9000 / sizeof(Readout)};
-
 private:
-  hdf5::file::File file_;
-  hdf5::datatype::Datatype dtype_;
-  hdf5::node::Dataset dataset_;
-  hdf5::dataspace::Hyperslab slab_{{0}, {chunk_size}};
+  static constexpr size_t u8 { sizeof(std::uint8_t)};
+  static constexpr size_t u16 { sizeof(std::uint16_t)};
+  static constexpr size_t u32 { sizeof(std::uint32_t)};
+  static constexpr size_t b { sizeof(bool)};
+
+public:
+  using Type = Readout;
+  using TypeClass = Compound;
+
+  static TypeClass create(const Type & = Type())
+  {
+    auto type = datatype::Compound::create(sizeof(Readout));
+    type.insert("fec",
+                0,
+                datatype::create<std::uint8_t>());
+    type.insert("chip_id",
+                u8,
+                datatype::create<std::uint8_t>());
+    type.insert("bonus_timestamp",
+                2*u8,
+                datatype::create<std::uint32_t>());
+    type.insert("srs_timestamp",
+                2*u8 + u32,
+                datatype::create<std::uint32_t>());
+    type.insert("channel",
+                2*u8 + 2*u32,
+                datatype::create<std::uint16_t>());
+    type.insert("bcid",
+                2*u8 + 2*u32 + u16,
+                datatype::create<std::uint16_t>());
+    type.insert("tdc",
+                2*u8 + 2*u32 + 2*u16,
+                datatype::create<std::uint16_t>());
+    type.insert("adc",
+                2*u8 + 2*u32 + 3*u16,
+                datatype::create<std::uint16_t>());
+    type.insert("over_threshold",
+                2*u8 + 2*u32 + 4*u16,
+                datatype::create<bool>());
+    return type;
+  }
 };
+}
+
+}
+
+using ReadoutFile = DumpFile<Readout>;
