+++ prototype2/gdgem/nmx/ReadoutFileTest.cpp
@@ -5,58 +5,82 @@
 
 class ReadoutFileTest : public TestBase {
 protected:
-  virtual void SetUp() {}
+  virtual void SetUp() {
+    if (boost::filesystem::exists("readout_file_test.h5"))
+    {
+      boost::filesystem::remove("readout_file_test.h5");
+    }
+
+    if (boost::filesystem::exists("readout_file_test_1.h5"))
+    {
+      boost::filesystem::remove("readout_file_test_1.h5");
+    }
+  }
   virtual void TearDown() {}
 };
 
 TEST_F(ReadoutFileTest, CreateFile) {
-  ReadoutFile::create("readout_file_test.h5");
+  ReadoutFile::create("readout_file_test");
   EXPECT_TRUE(hdf5::file::is_hdf5_file("readout_file_test.h5"));
 }
 
 TEST_F(ReadoutFileTest, OpenEmptyFile) {
-  ReadoutFile::create("readout_file_test.h5");
-  auto file = ReadoutFile::open("readout_file_test.h5");
-  EXPECT_EQ(file.count(), 0);
+  ReadoutFile::create("readout_file_test");
+  auto file = ReadoutFile::open("readout_file_test");
+  EXPECT_EQ(file->count(), 0);
 }
 
-TEST_F(ReadoutFileTest, Write) {
-  auto file = ReadoutFile::create("readout_file_test.h5");
-  file.data.resize(3);
-  file.write();
-  EXPECT_EQ(file.count(), 3);
-  file.write();
-  EXPECT_EQ(file.count(), 6);
-  file.write();
-  EXPECT_EQ(file.count(), 9);
+TEST_F(ReadoutFileTest, Push) {
+  auto file = ReadoutFile::create("readout_file_test");
+  file->push(std::vector<Readout>(100, Readout()));
+  EXPECT_EQ(file->count(), 0);
+  file->push(std::vector<Readout>(900, Readout()));
+  EXPECT_EQ(file->count(), 1000);
+  file->push(std::vector<Readout>(3000, Readout()));
+  EXPECT_EQ(file->count(), 4000);
+}
+
+TEST_F(ReadoutFileTest, PushFileRotation) {
+  auto file = ReadoutFile::create("readout_file_test", 1);
+  file->push(std::vector<Readout>(100, Readout()));
+  EXPECT_EQ(file->count(), 0);
+  file->push(std::vector<Readout>(900, Readout()));
+  EXPECT_EQ(file->count(), 1000);
+  file->push(std::vector<Readout>(3000, Readout()));
+  EXPECT_EQ(file->count(), 4000);
+
+  EXPECT_TRUE(hdf5::file::is_hdf5_file("readout_file_test.h5"));
+
+  EXPECT_FALSE(boost::filesystem::exists("readout_file_test_1.h5"));
+  file->push(std::vector<Readout>(300000, Readout()));
+  EXPECT_TRUE(hdf5::file::is_hdf5_file("readout_file_test_1.h5"));
 }
 
 TEST_F(ReadoutFileTest, Read) {
-  auto file_out = ReadoutFile::create("readout_file_test.h5");
-  file_out.data.resize(9);
-  file_out.write();
-  file_out = ReadoutFile();
+  auto file_out = ReadoutFile::create("readout_file_test");
+  file_out->push(std::vector<Readout>(900, Readout()));
+  file_out.reset();
 
-  auto file = ReadoutFile::open("readout_file_test.h5");
-  EXPECT_EQ(file.data.size(), 0);
-  file.read_at(0, 3);
-  EXPECT_EQ(file.data.size(), 3);
-  file.read_at(0, 9);
-  EXPECT_EQ(file.data.size(), 9);
+  auto file = ReadoutFile::open("readout_file_test");
+  EXPECT_EQ(file->Data.size(), 0);
+  file->readAt(0, 3);
+  EXPECT_EQ(file->Data.size(), 3);
+  file->readAt(0, 9);
+  EXPECT_EQ(file->Data.size(), 9);
 
-  EXPECT_THROW(file.read_at(0, 12), std::runtime_error);
+  EXPECT_THROW(file->readAt(0, 1000), std::runtime_error);
 }
 
 TEST_F(ReadoutFileTest, ReadAll) {
-  auto file_out = ReadoutFile::create("readout_file_test.h5");
-  file_out.data.resize(9);
-  file_out.write();
-  file_out = ReadoutFile();
+  auto file_out = ReadoutFile::create("readout_file_test");
+  file_out->push(std::vector<Readout>(900, Readout()));
+  file_out.reset();
 
   std::vector<Readout> data;
-  ReadoutFile::read("readout_file_test.h5", data);
-  EXPECT_EQ(data.size(), 9);
+  ReadoutFile::read("readout_file_test", data);
+  EXPECT_EQ(data.size(), 900);
 }
+
 
 int main(int argc, char **argv) {
   testing::InitGoogleTest(&argc, argv);
