+++ prototype2/multigrid/mgmesytec/Hit.h
@@ -29,6 +29,10 @@ struct __attribute__ ((packed)) Hit {
 
   // \brief prints values for debug purposes
   std::string debug() const;
+
+  // \todo use constexpr string_view when c++17 arrives
+  static std::string DatasetName() { return "mgmesytec_hits"; }
+  static std::string FormatVersion() { return "1.0.0"; }
 };
 
 }
